Moved testCoreJavaParity from TestRegExpQuery to TestRegExp
