+++ lucene/core/src/test/org/apache/lucene/util/automaton/TestRegExp.java
@@ -17,7 +17,11 @@
 package org.apache.lucene.util.automaton;
 
 
+import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.LuceneTestCase;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class TestRegExp extends LuceneTestCase {
 
@@ -83,4 +87,127 @@ public class TestRegExp extends LuceneTestCase {
     a = new RegExp("#?").toAutomaton(1000);
     assertTrue(a.toString().length() > 0);
   }
+  
+  public void testCoreJavaParity() {
+    // Generate random doc values and random regular expressions
+    // and check for same matching behaviour as Java's Pattern class.
+    for (int i = 0; i < 1000; i++) {
+      checkRandomExpression(randomDocValue(1 + random().nextInt(30)));
+    }        
+  }
+
+  static String randomDocValue(int minLength) {
+    String charPalette = "AAAaaaBbbCccc123456 \t";
+    StringBuilder sb = new StringBuilder();
+    for (int i = 0; i < minLength; i++) {
+      sb.append(charPalette.charAt(randomInt(charPalette.length() - 1)));
+    }
+    return sb.toString();
+  }
+
+  private static int randomInt(int bound) {
+    return bound == 0 ? 0 : random().nextInt(bound);
+  }
+
+  protected String checkRandomExpression(String docValue) {
+    // Generate and test a random regular expression which should match the given docValue
+    StringBuilder result = new StringBuilder();
+    // Pick a part of the string to change
+    int substitutionPoint = randomInt(docValue.length() - 1);
+    int substitutionLength = 1 + randomInt(Math.min(10, docValue.length() - substitutionPoint));
+
+    // Add any head to the result, unchanged
+    if (substitutionPoint > 0) {
+      result.append(docValue.substring(0, substitutionPoint));
+    }
+
+    // Modify the middle...
+    String replacementPart = docValue.substring(substitutionPoint, substitutionPoint + substitutionLength);
+    int mutation = random().nextInt(13);
+    switch (mutation) {
+      case 0:
+        // OR with random alpha of same length
+        result.append("(" + replacementPart + "|d" + randomDocValue(replacementPart.length()) + ")");
+        break;
+      case 1:
+        // OR with non-existant value
+        result.append("(" + replacementPart + "|doesnotexist)");
+        break;
+      case 2:
+        // OR with another randomised regex (used to create nested levels of expression).
+        result.append("(" + checkRandomExpression(replacementPart) + "|doesnotexist)");
+        break;
+      case 3:
+        // Star-replace all ab sequences.
+        result.append(replacementPart.replaceAll("ab", ".*"));
+        break;
+      case 4:
+        // .-replace all b chars
+        result.append(replacementPart.replaceAll("b", "."));
+        break;
+      case 5:
+        // length-limited stars {1,2}
+        result.append(".{1," + replacementPart.length() + "}");
+        break;
+      case 6:
+        // replace all chars with .
+        result.append(replacementPart.replaceAll(".", "."));
+        break;
+      case 7:
+        // OR with uppercase chars eg [aA] (many of these sorts of expression in the wild..
+        char[] chars = replacementPart.toCharArray();
+        for (char c : chars) {
+          result.append("[" + c + Character.toUpperCase(c) + "]");
+        }
+        break;
+      case 8:
+        // NOT a character - replace all b's with "not a"
+        result.append(replacementPart.replaceAll("b", "[^a]"));
+        break;
+      case 9:
+        // Make whole part repeatable 1 or more times
+        result.append("(" + replacementPart + ")+");
+        break;
+      case 10:
+        // Make whole part repeatable 0 or more times
+        result.append("(" + replacementPart + ")?");
+        break;
+      case 11:
+        // Make any digits replaced by character class
+        result.append(replacementPart.replaceAll("\\d", "\\\\d"));
+        break;
+      case 12:
+        // Make any whitespace chars replaced by not word class
+        result.append(replacementPart.replaceAll("\\s", "\\\\W"));
+        break;
+      case 13:
+        // Make any whitespace chars replace by whitespace class
+        result.append(replacementPart.replaceAll("\\s", "\\\\s"));
+        break;
+      default:
+        break;
+    }
+    // add any remaining tail, unchanged
+    if (substitutionPoint + substitutionLength <= docValue.length() - 1) {
+      result.append(docValue.substring(substitutionPoint + substitutionLength));
+    }
+
+    String regexPattern = result.toString();
+    // Assert our randomly generated regex actually matches the provided raw input using java's expression matcher
+    Pattern pattern = Pattern.compile(regexPattern);
+    Matcher matcher = pattern.matcher(docValue);
+    assertTrue("Java regex " + regexPattern + " did not match doc value " + docValue, matcher.matches());
+
+    RegExp regex = new RegExp(regexPattern);
+    Automaton automaton = regex.toAutomaton();
+    ByteRunAutomaton bytesMatcher = new ByteRunAutomaton(automaton);
+    BytesRef br = new BytesRef(docValue);
+    assertTrue(
+        "[" + regexPattern + "]should match [" + docValue + "]" + substitutionPoint + "-" + substitutionLength + "/"
+            + docValue.length(),
+        bytesMatcher.run(br.bytes, br.offset, br.length)
+    );
+    return regexPattern;
+  }
+  
 }
